requirements.txt
streamlit
tensorflow
huggingface_hub
numpy
pillow
matplotlib
streamlit-drawable-canvas
